Oly empty MCBTree can be removed to make deletion of indexes durable operation

+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/multivalue/v1/OCellBTreeMultiValueV1.java
@@ -500,6 +500,10 @@ public final class OCellBTreeMultiValueV1<K> extends ODurableComponent implement
     try {
       acquireExclusiveLock();
       try {
+        if (size() > 0) {
+          throw new OCellBTreeMultiValueException("Only empty tries can be removed", this);
+        }
+
         deleteFile(fileId);
         deleteFile(nullBucketFileId);
       } finally {
